+++ src/main/java/com/ubiqsecurity/LoadSearchKeys.java
@@ -9,6 +9,11 @@ import com.google.gson.*;
 import com.ubiqsecurity.structured.FF1;
 import java.util.Base64;
 import java.security.SecureRandom;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
+
+import java.util.ArrayList;
 
 class LoadSearchKeys  {
     private static boolean verbose= false;
@@ -25,122 +30,188 @@ class LoadSearchKeys  {
       return sb.toString();
    }
   
-
+    // Reset TTL means that if a Key already exists, then execute a Put(Get) to reset the cache TTL
+    // This is a specific use case for UpdateCache
     public static void loadKeys(
-      UbiqCredentials ubiqCredentials,
       UbiqWebServices ubiqWebServices,
       FFS ffs,
       FFXCache ffxCache,
-      String ffs_name) throws Exception {
+      String[] datasets) throws Exception {
       String csu = "loadKeys";
+      List<String> datasetsToRemove = new ArrayList<>();
 
+      Boolean verbose = false;
       // Call the web services to get the search keys
 
-      if (verbose) System.out.println(String.format("%s started  \n", csu));
+      if (verbose) System.out.println(String.format("%s started", csu));
 
       // Already swaps out the encrypted_private_key if this is IDP
-      JsonObject fpe_search_keys = ubiqWebServices.getFpeDefKeys(ffs_name);
+      JsonObject fpe_search_keys = ubiqWebServices.getFpeDefKeys(datasets);
 
-      if (verbose) System.out.println(String.format("%s before top_level  \n", csu));
+      if (verbose) System.out.println(String.format("%s before top_level", csu));
 
-      JsonObject top_level = fpe_search_keys.get(ffs_name).getAsJsonObject();
-
-      if (verbose) System.out.println(String.format("%s before dataset  \n", csu));
-      
-      JsonObject dataset = top_level.get("ffs").getAsJsonObject();
-
-      // If Dataset (FFS) is not already in the FFS Cache, add it.
-
-     FFS_Record ffsRecord = FFS_Record.parse(dataset);
-
-      if (verbose) System.out.println(String.format("%s ffsRecord %s  \n", csu, (new Gson()).toJson(ffsRecord)));
-
-      if (verbose) System.out.println(String.format("%s tweak %s  \n", csu, ffsRecord.getTweak()));
-      if (!ffs.FFSCache.asMap().containsKey(ffs_name)) {
-        if (verbose) System.out.println(String.format("%s FFSCache miss %s  \n", csu, ffs_name));
-        ffs.FFSCache.put(ffs_name, ffsRecord);
-      } else {
-        if (verbose) System.out.println(String.format("%s FFSCache HIT %s  \n", csu, ffs_name));
-      }
-
-      if (verbose) System.out.println(String.format("%s before encrypted_private_key  \n", csu));
-      String encrypted_private_key = top_level.get("encrypted_private_key").getAsString();
-
-      // if (ubiqCredentials.isIdp()) {
-      //   encrypted_private_key = ubiqCredentials.getEncryptedPrivateKey();
-      // }
-      
-      if (verbose) System.out.println(String.format("%s encrypted_private_key  %s\n", csu, encrypted_private_key));
-
-      Integer current_key_number =top_level.get("current_key_number").getAsInt();
-      if (verbose) System.out.println(String.format("%s current_key_number  %d\n", csu, current_key_number));
-
-      JsonArray keys = top_level.get("keys").getAsJsonArray();
-
-      if (verbose) System.out.println(String.format("%s arrayCount  %d\n", csu, keys.size()));
-      
-      // Loop over the keys.  If not alraedy
-
-      for (int i = 0; i < keys.size(); i++) {
-        byte[] tweak = null;
-
-        FFS_KeyId keyId = new FFS_KeyId(ffsRecord, i);
-
-        if (!ffxCache.FFXCache.asMap().containsKey(keyId)) {
-          if (verbose) System.out.println(String.format("%s  FFXCache does not contain key for %d\n", csu, i));
-          byte[] key = ubiqWebServices.getUnwrappedKey(encrypted_private_key, keys.get(i).getAsString());
-
-          if (verbose) System.out.println(String.format("%s byte[] key %d\n", csu, key.length));
-
-          FFX_Ctx ctx = new FFX_Ctx();
-          if (keyId.ffs.getTweakSource().equals("constant")) {
-            if (verbose) System.out.println(String.format("%s  tweak source %s\n", csu,keyId.ffs.getTweakSource()));
-            String s = keyId.ffs.getTweak();
-            if (verbose) System.out.println(String.format("%s  tweak  %s\n", csu, s));
-
-            if (verbose) System.out.println(String.format("%s getting tweak %d\n", csu, keyId.ffs.getTweak().length()));
-            tweak= Base64.getDecoder().decode(keyId.ffs.getTweak());
-            if (verbose) System.out.println(String.format("%s byte[] tweak %d\n", csu, tweak.length));
+      // If there were not dataset names, then all compatible datasets were returned.  Permissions
+      // may have been made, so need to invalidate the datasets cache, not yet invalidating the keys.
+      if (datasets.length == 0) {
+        // Pretend like all datasets were passed in
+        datasets = fpe_search_keys.keySet().toArray(new String[0]);
+        
+        if (verbose) {
+          System.out.println(String.format("%s Found Datasets:", csu));
+          for (String s : datasets) {
+            System.out.println(String.format("\t %s", s));
           }
+        }
+        // Get a list of all the existing cached dataset names since one of the 
+        // existing caches might need to be removed if it doesn't match the 
+        // name returned by the web call.
+        Set<String> existingCachedNames = ffs.FFSCache.asMap().keySet();
+        if (verbose) {
+          System.out.println(String.format("%s Existing Cached Datasets:", csu));
+          for (String s : existingCachedNames) {
+            System.out.println(String.format("\t %s", s));
+          }
+        }
+        // For all the existing cached names, exclude the datasets returned, since those will be added / maintained
+        existingCachedNames.removeAll(Arrays.asList(datasets));
+        datasetsToRemove = new ArrayList<>(existingCachedNames);
+        if (verbose) {
+          System.out.println(String.format("%s Cached Datsets to remove:", csu));
+          for (String s : datasetsToRemove) {
+            System.out.println(String.format("\t %s", s));
+          }
+        }
+      } // All datasets retrieved
 
-          if (verbose) System.out.println(String.format("%s FFX_Ctx ctx %s\n", csu, "after"));
+      // For all the datasets retrieved, refresh the cache.
+      for (String dataset_name : datasets)
+      {
+        JsonElement element = fpe_search_keys.get(dataset_name);
+        // It is possible that the requested dataset does not exist in the results.
+        // If the case of all datasets, this is always false.
+        if (element == null || element.isJsonNull())
+        {
+          if (verbose) System.out.println(String.format("%s dataset was cached but will now be removed %s", csu, dataset_name));
+          datasetsToRemove.add(dataset_name);
+          continue;
+        }
+        JsonObject top_level = element.getAsJsonObject();
 
-          ctx.setFF1(new FF1(key, tweak, 
-                keyId.ffs.getMinTweakLength(), 
-                keyId.ffs.getMaxTweakLength(), 
-                keyId.ffs.getInputCharacterSet().length(), keyId.ffs.getInputCharacterSet()), 
-                i);
+        if (verbose) System.out.println(String.format("%s before dataset", csu));
+        
+        JsonObject dataset = top_level.get("ffs").getAsJsonObject();
 
-          ffxCache.FFXCache.put(keyId, ctx);
-          if (verbose) System.out.println(String.format("%s ffxCache.FFXCache.put(keyId, ctx); %s\n", csu, "after"));
+        FFS_Record ffsRecord = FFS_Record.parse(dataset);
 
-          if (i == current_key_number) {
-            FFS_KeyId currentKey = new FFS_KeyId(ffsRecord, null);
-            if (!ffxCache.FFXCache.asMap().containsKey(currentKey)) {
-              if (verbose) System.out.println(String.format("%s  FFXCache does not contain current_key_number %d\n", csu, current_key_number));
+        if (verbose) System.out.println(String.format("%s ffsRecord %s ", csu, (new Gson()).toJson(ffsRecord)));
 
-              FFX_Ctx ctx2 = new FFX_Ctx();
-              if (currentKey.ffs.getTweakSource().equals("constant")) {
-                tweak= Base64.getDecoder().decode(currentKey.ffs.getTweak());
-              }
-    
-              ctx2.setFF1(new FF1(key, tweak, 
-                    currentKey.ffs.getMinTweakLength(), 
-                    currentKey.ffs.getMaxTweakLength(), 
-                    currentKey.ffs.getInputCharacterSet().length(), currentKey.ffs.getInputCharacterSet()), 
-                    i);
+        if (verbose) System.out.println(String.format("%s tweak %s", csu, ffsRecord.getTweak()));
+        // Permissions may have changed, so overwrite the existing dataset definition, even if it already exists
+        ffs.FFSCache.put(dataset_name, ffsRecord);
 
-              ffxCache.FFXCache.put(currentKey, ctx2);
-            } else {
-              if (verbose) System.out.println(String.format("%s  FFXCache contains current_key_number %d\n", csu, current_key_number));
+        if (verbose) System.out.println(String.format("%s before encrypted_private_key", csu));
+        String encrypted_private_key = top_level.get("encrypted_private_key").getAsString();
 
+        if (verbose) System.out.println(String.format("%s encrypted_private_key  %s", csu, encrypted_private_key));
+
+        Integer current_key_number =top_level.get("current_key_number").getAsInt();
+        if (verbose) System.out.println(String.format("%s current_key_number  %d", csu, current_key_number));
+
+        JsonArray keys = top_level.get("keys").getAsJsonArray();
+
+        if (verbose) System.out.println(String.format("%s arrayCount  %d", csu, keys.size()));
+        
+        // Loop over the keys.  If not already
+
+        for (int i = 0; i < keys.size(); i++) {
+          byte[] tweak = null;
+
+          FFS_KeyId keyId = new FFS_KeyId(ffsRecord, i);
+
+          if (!ffxCache.FFXCache.asMap().containsKey(keyId)) {
+            if (verbose) System.out.println(String.format("%s  FFXCache does not contain key for %d", csu, i));
+            byte[] key = ubiqWebServices.getUnwrappedKey(encrypted_private_key, keys.get(i).getAsString());
+
+            if (verbose) System.out.println(String.format("%s byte[] key %d", csu, key.length));
+
+            FFX_Ctx ctx = new FFX_Ctx();
+            if (keyId.ffs.getTweakSource().equals("constant")) {
+              if (verbose) System.out.println(String.format("%s  tweak source %s", csu,keyId.ffs.getTweakSource()));
+              String s = keyId.ffs.getTweak();
+              if (verbose) System.out.println(String.format("%s  tweak  %s", csu, s));
+
+              if (verbose) System.out.println(String.format("%s getting tweak %d", csu, keyId.ffs.getTweak().length()));
+              tweak= Base64.getDecoder().decode(keyId.ffs.getTweak());
+              if (verbose) System.out.println(String.format("%s byte[] tweak %d", csu, tweak.length));
             }
-          }
-        } else {
-          if (verbose) System.out.println(String.format("%s  FFXCache already exists for %d\n", csu, i));
 
+            if (verbose) System.out.println(String.format("%s FFX_Ctx ctx %s", csu, "after"));
+
+            ctx.setFF1(new FF1(key, tweak, 
+                  keyId.ffs.getMinTweakLength(), 
+                  keyId.ffs.getMaxTweakLength(), 
+                  keyId.ffs.getInputCharacterSet().length(), keyId.ffs.getInputCharacterSet()), 
+                  i);
+
+            ffxCache.FFXCache.put(keyId, ctx);
+            if (verbose) System.out.println(String.format("%s ffxCache.FFXCache.put(keyId, ctx); %s", csu, "after"));
+
+            if (i == current_key_number) {
+              FFS_KeyId currentKey = new FFS_KeyId(ffsRecord, null);
+              if (!ffxCache.FFXCache.asMap().containsKey(currentKey)) {
+                if (verbose) System.out.println(String.format("%s  FFXCache does not contain current_key_number %d", csu, current_key_number));
+
+                FFX_Ctx ctx2 = new FFX_Ctx();
+                if (currentKey.ffs.getTweakSource().equals("constant")) {
+                  tweak= Base64.getDecoder().decode(currentKey.ffs.getTweak());
+                }
+      
+                ctx2.setFF1(new FF1(key, tweak, 
+                      currentKey.ffs.getMinTweakLength(), 
+                      currentKey.ffs.getMaxTweakLength(), 
+                      currentKey.ffs.getInputCharacterSet().length(), currentKey.ffs.getInputCharacterSet()), 
+                      i);
+
+                ffxCache.FFXCache.put(currentKey, ctx2);
+              } else {
+                if (verbose) System.out.println(String.format("%s  FFXCache contains current_key_number %d", csu, current_key_number));
+                ffxCache.FFXCache.put(currentKey, ffxCache.FFXCache.get(currentKey));
+              }
+            }
+          } else {
+            if (verbose) System.out.println(String.format("%s  FFXCache already exists for %d", csu, i));
+            ffxCache.FFXCache.put(keyId, ffxCache.FFXCache.get(keyId));
+          }
         }
       }
+
+      // Requested datasets that were not found are in the datasetsToRemove to remove list
+
+      // In the case where datasets was empty, request all datasets, the datasets that were loaded but no longer in
+      // the datasets retrieved will be in the datasetsToRemove list.
+      
+      for (String dataset_name : datasetsToRemove)
+      {
+        if (verbose) System.out.println("Dataset to remove: " + dataset_name);
+        if (verbose) System.out.println(String.format("%s  Checking FFS Cache for %s", csu, dataset_name));
+
+        if (ffs.FFSCache.asMap().containsKey(dataset_name)) {
+          if (verbose) System.out.println(String.format("%s  Found %s in FFS Cache", csu, dataset_name));
+          FFS_Record ffsRecord = ffs.FFSCache.get(dataset_name);
+        
+          FFS_KeyId currentKey = new FFS_KeyId(ffsRecord, null);
+          if (ffxCache.FFXCache.asMap().containsKey(currentKey)) {
+            FFX_Ctx currentCtx = ffxCache.FFXCache.get(currentKey);
+            if (verbose) System.out.println(String.format("%s  Found current key (%d) in FFXCache ", csu, currentCtx.getKeyNumber()));
+            for (Integer i = 0; i < currentCtx.getKeyNumber(); i++) {
+              ffxCache.FFXCache.invalidate(new FFS_KeyId(ffsRecord, i));
+            }
+            ffxCache.FFXCache.invalidate(currentKey);
+            if (verbose) System.out.println(String.format("%s  Invalidating keys (%d) - (%d) in FFXCache ", csu, 0, currentCtx.getKeyNumber()));
+          }
+        }
+      }
+      ffs.FFSCache.invalidateAll(datasetsToRemove);
     }
 
     // Used to explicitly load data into the cache from something such as APIGEE
@@ -154,15 +225,15 @@ class LoadSearchKeys  {
 
       // Call the web services to get the search keys
 
-      if (verbose) System.out.println(String.format("%s started  \n", csu));
+      if (verbose) System.out.println(String.format("%s started  ", csu));
 
       // JsonObject fpe_search_keys = ubiqWebServices.getFpeDefKeys(ffs_name);
 
-      if (verbose) System.out.println(String.format("%s before top_level  \n", csu));
+      if (verbose) System.out.println(String.format("%s before top_level  ", csu));
 
       // JsonObject top_level = fpe_search_keys.get(dataset_name).getAsJsonObject();
 
-      if (verbose) System.out.println(String.format("%s before dataset  \n", csu));
+      if (verbose) System.out.println(String.format("%s before dataset  ", csu));
       
       JsonObject dataset = fpe_search_keys.get("ffs").getAsJsonObject();
       String dataset_name = dataset.get("name").getAsString();
@@ -171,27 +242,27 @@ class LoadSearchKeys  {
 
       FFS_Record ffsRecord = FFS_Record.parse(dataset);
 
-      if (verbose) System.out.println(String.format("%s ffsRecord %s  \n", csu, (new Gson()).toJson(ffsRecord)));
+      if (verbose) System.out.println(String.format("%s ffsRecord %s  ", csu, (new Gson()).toJson(ffsRecord)));
 
-      if (verbose) System.out.println(String.format("%s tweak %s  \n", csu, ffsRecord.getTweak()));
+      if (verbose) System.out.println(String.format("%s tweak %s  ", csu, ffsRecord.getTweak()));
       if (!ffs.FFSCache.asMap().containsKey(dataset_name)) {
-        if (verbose) System.out.println(String.format("%s FFSCache miss %s  \n", csu, dataset_name));
+        if (verbose) System.out.println(String.format("%s FFSCache miss %s  ", csu, dataset_name));
         ffs.FFSCache.put(dataset_name, ffsRecord);
       } else {
-        if (verbose) System.out.println(String.format("%s FFSCache HIT %s  \n", csu, dataset_name));
+        if (verbose) System.out.println(String.format("%s FFSCache HIT %s  ", csu, dataset_name));
       }
 
-      if (verbose) System.out.println(String.format("%s before encrypted_private_key  \n", csu));
+      if (verbose) System.out.println(String.format("%s before encrypted_private_key  ", csu));
       String encrypted_private_key = fpe_search_keys.get("encrypted_private_key").getAsString();
       
-      if (verbose) System.out.println(String.format("%s encrypted_private_key  %s\n", csu, encrypted_private_key));
+      if (verbose) System.out.println(String.format("%s encrypted_private_key  %s", csu, encrypted_private_key));
 
       Integer current_key_number =fpe_search_keys.get("current_key_number").getAsInt();
-      if (verbose) System.out.println(String.format("%s current_key_number  %d\n", csu, current_key_number));
+      if (verbose) System.out.println(String.format("%s current_key_number  %d", csu, current_key_number));
 
       JsonArray keys = fpe_search_keys.get("keys").getAsJsonArray();
 
-      if (verbose) System.out.println(String.format("%s arrayCount  %d\n", csu, keys.size()));
+      if (verbose) System.out.println(String.format("%s arrayCount  %d", csu, keys.size()));
       
       // Loop over the keys.  If not alraedy
 
@@ -201,23 +272,23 @@ class LoadSearchKeys  {
         FFS_KeyId keyId = new FFS_KeyId(ffsRecord, i);
 
         if (!ffxCache.FFXCache.asMap().containsKey(keyId)) {
-          if (verbose) System.out.println(String.format("%s  FFXCache does not contain key for %d\n", csu, i));
+          if (verbose) System.out.println(String.format("%s  FFXCache does not contain key for %d", csu, i));
           byte[] key = ubiqWebServices.getUnwrappedKey(encrypted_private_key, keys.get(i).getAsString());
 
-          if (verbose) System.out.println(String.format("%s byte[] key %d\n", csu, key.length));
+          if (verbose) System.out.println(String.format("%s byte[] key %d", csu, key.length));
 
           FFX_Ctx ctx = new FFX_Ctx();
           if (keyId.ffs.getTweakSource().equals("constant")) {
-            if (verbose) System.out.println(String.format("%s  tweak source %s\n", csu,keyId.ffs.getTweakSource()));
+            if (verbose) System.out.println(String.format("%s  tweak source %s", csu,keyId.ffs.getTweakSource()));
             String s = keyId.ffs.getTweak();
-            if (verbose) System.out.println(String.format("%s  tweak  %s\n", csu, s));
+            if (verbose) System.out.println(String.format("%s  tweak  %s", csu, s));
 
-            if (verbose) System.out.println(String.format("%s getting tweak %d\n", csu, keyId.ffs.getTweak().length()));
+            if (verbose) System.out.println(String.format("%s getting tweak %d", csu, keyId.ffs.getTweak().length()));
             tweak= Base64.getDecoder().decode(keyId.ffs.getTweak());
-            if (verbose) System.out.println(String.format("%s byte[] tweak %d\n", csu, tweak.length));
+            if (verbose) System.out.println(String.format("%s byte[] tweak %d", csu, tweak.length));
           }
 
-          if (verbose) System.out.println(String.format("%s FFX_Ctx ctx %s\n", csu, "after"));
+          if (verbose) System.out.println(String.format("%s FFX_Ctx ctx %s", csu, "after"));
 
           ctx.setFF1(new FF1(key, tweak, 
                 keyId.ffs.getMinTweakLength(), 
@@ -226,12 +297,12 @@ class LoadSearchKeys  {
                 i);
 
           ffxCache.FFXCache.put(keyId, ctx);
-          if (verbose) System.out.println(String.format("%s ffxCache.FFXCache.put(keyId, ctx); %s\n", csu, "after"));
+          if (verbose) System.out.println(String.format("%s ffxCache.FFXCache.put(keyId, ctx); %s", csu, "after"));
 
           if (i == current_key_number) {
             FFS_KeyId currentKey = new FFS_KeyId(ffsRecord, null);
             if (!ffxCache.FFXCache.asMap().containsKey(currentKey)) {
-              if (verbose) System.out.println(String.format("%s  FFXCache does not contain current_key_number %d\n", csu, current_key_number));
+              if (verbose) System.out.println(String.format("%s  FFXCache does not contain current_key_number %d", csu, current_key_number));
 
               FFX_Ctx ctx2 = new FFX_Ctx();
               if (currentKey.ffs.getTweakSource().equals("constant")) {
@@ -246,12 +317,12 @@ class LoadSearchKeys  {
 
               ffxCache.FFXCache.put(currentKey, ctx2);
             } else {
-              if (verbose) System.out.println(String.format("%s  FFXCache contains current_key_number %d\n", csu, current_key_number));
+              if (verbose) System.out.println(String.format("%s  FFXCache contains current_key_number %d", csu, current_key_number));
 
             }
           }
         } else {
-          if (verbose) System.out.println(String.format("%s  FFXCache already exists for %d\n", csu, i));
+          if (verbose) System.out.println(String.format("%s  FFXCache already exists for %d", csu, i));
 
         }
       }
@@ -268,12 +339,12 @@ class LoadSearchKeys  {
         String dataset_name = dataset.get("name").getAsString();
 
         if (!ffs.FFSCache.asMap().containsKey(dataset_name)) {
-          if (verbose) System.out.println(String.format("%s FFSCache miss %s  \n", csu, dataset_name));
+          if (verbose) System.out.println(String.format("%s FFSCache miss %s  ", csu, dataset_name));
 
           FFS_Record ffsRecord = FFS_Record.parse(dataset);
           ffs.FFSCache.put(dataset_name, ffsRecord);
         } else {
-          if (verbose) System.out.println(String.format("%s FFSCache HIT %s  \n", csu, dataset_name));
+          if (verbose) System.out.println(String.format("%s FFSCache HIT %s  ", csu, dataset_name));
         }
         return dataset_name;
     }
@@ -353,9 +424,9 @@ class LoadSearchKeys  {
         if (!ffxCache.FFXCache.asMap().containsKey(keyId)) {
           FFX_Ctx ctx= makeCtx(ffsRecord, key, key_number);
           ffxCache.FFXCache.put(keyId, ctx);
-          if (verbose) System.out.println(String.format("%s FFXCache MISS %s %d \n", csu, dataset_name, key_number));
+          if (verbose) System.out.println(String.format("%s FFXCache MISS %s %d ", csu, dataset_name, key_number));
         } else {
-          if (verbose) System.out.println(String.format("%s FFXCache HIT %s %d \n", csu, dataset_name, key_number));
+          if (verbose) System.out.println(String.format("%s FFXCache HIT %s %d ", csu, dataset_name, key_number));
         }
         // For Encrypt
         if (current_key_flag) {
@@ -363,9 +434,9 @@ class LoadSearchKeys  {
           if (!ffxCache.FFXCache.asMap().containsKey(nullKeyId)) {
             FFX_Ctx ctx= makeCtx(ffsRecord, key, key_number);
             ffxCache.FFXCache.put(nullKeyId, ctx);
-            if (verbose) System.out.println(String.format("%s FFXCache MISS %s %d \n", csu, dataset_name, key_number));
+            if (verbose) System.out.println(String.format("%s FFXCache MISS %s %d ", csu, dataset_name, key_number));
           } else {
-            if (verbose) System.out.println(String.format("%s FFXCache HIT %s %d \n", csu, dataset_name, key_number));
+            if (verbose) System.out.println(String.format("%s FFXCache HIT %s %d ", csu, dataset_name, key_number));
           }
         }
     }
@@ -389,7 +460,7 @@ class LoadSearchKeys  {
       try {
 
         // System.out.println("data: " + data);
-        System.out.println("encryption_key: " + encryption_key);
+        if (verbose) System.out.println("encryption_key: " + encryption_key);
 
         AlgorithmInfo alg = new AlgorithmInfo(data.get("alg").getAsString());
         byte[] initVector = Base64.getDecoder().decode(data.get("iv").getAsString());
@@ -397,7 +468,7 @@ class LoadSearchKeys  {
         byte[] key = Base64.getDecoder().decode(encryption_key);
         byte[] empty = null;
 
-        System.out.println("alg: " + data.get("alg").getAsString());
+        if (verbose) System.out.println("alg: " + data.get("alg").getAsString());
 
         if (key.length > alg.getKeyLength()) {
           byte[] tmp = new byte[alg.getKeyLength()];
@@ -440,7 +511,7 @@ class LoadSearchKeys  {
       JsonObject results = new JsonObject();
       try {
 
-      System.out.println("data: " + data_bytes.toString() + "  encryption_key: " + encryption_key);
+      if (verbose) System.out.println("data: " + data_bytes.toString() + "  encryption_key: " + encryption_key);
       AlgorithmInfo alg = new AlgorithmInfo(algorithm_name);
       // byte[] data_bytes = data.getBytes();
       byte[] encrypted_data;
@@ -453,9 +524,9 @@ class LoadSearchKeys  {
 
       // Will be 33 characters.  Limit to alg.getKeyLength
       byte[] key = Base64.getDecoder().decode(encryption_key);
-      System.out.println("tmp.length: " + key.length);
+      if (verbose) System.out.println("tmp.length: " + key.length);
 
-      System.out.println("secretCryptoAccessKey.toCharArray:" + encryption_key.toCharArray().length);
+      if (verbose) System.out.println("secretCryptoAccessKey.toCharArray:" + encryption_key.toCharArray().length);
 
       if (key.length > alg.getKeyLength()) {
         byte[] tmp = new byte[alg.getKeyLength()];